requirements.txt
paho-mqtt
flask
pyyaml
influxdb-client